+++ CHANGELOG.md
@@ -14,6 +14,7 @@ and to the [CHANGELOG recommendations](http://keepachangelog.com/).
 
 ### Fixed
 - Remove error when booking duration is less than 1 hour in booking price form
+- Remove errors fields message on new booking page when a secondary submission (Voucher, Delivery, ...) is done
 
 ### Changed
 - Disable 301 redirect when offerer go to the new booking page of its listing